+++ cliente/modelo/DatosTienda.h
@@ -7,8 +7,8 @@
 class DatosTienda{
 public:
 	std::atomic<bool> activo;
-	std::atomic<int> inventario[INV_ANCHO*INV_ALTO];
-	std::atomic<int> precios[INV_ANCHO*INV_ALTO];
+	std::atomic<uint16_t> inventario[INV_ANCHO*INV_ALTO];
+	std::atomic<uint16_t> precios[INV_ANCHO*INV_ALTO];
 	DatosTienda();
 };
